# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Conferência de Pedidos API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "operationId": "me_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUser"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/auth/resumo-dia": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Resumo Dia",
        "operationId": "resumo_dia_api_auth_resumo_dia_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/auth/perfil": {
      "put": {
        "tags": [
          "auth"
        ],
        "summary": "Atualizar Perfil",
        "operationId": "atualizar_perfil_api_auth_perfil_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PerfilRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUser"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/auth/senha": {
      "put": {
        "tags": [
          "auth"
        ],
        "summary": "Alterar Senha",
        "operationId": "alterar_senha_api_auth_senha_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SenhaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/auth/recuperar-senha": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Recuperar Senha",
        "operationId": "recuperar_senha_api_auth_recuperar_senha_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecuperarSenhaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/redefinir-senha": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Redefinir Senha",
        "operationId": "redefinir_senha_api_auth_redefinir_senha_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedefinirSenhaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "operationId": "logout_api_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthUser": {
        "properties": {
          "usuario": {
            "type": "string",
            "title": "Usuario"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "imagem": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imagem"
          },
          "telas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Telas",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "usuario",
          "nome",
          "role"
        ],
        "title": "AuthUser"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "usuario": {
            "type": "string",
            "title": "Usuario"
          },
          "senha": {
            "type": "string",
            "title": "Senha"
          }
        },
        "type": "object",
        "required": [
          "usuario",
          "senha"
        ],
        "title": "LoginRequest"
      },
      "LoginResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "LoginResponse"
      },
...
      "PerfilRequest": {
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Nome"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 254
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "imagem": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 600000
              },
              {
                "type": "null"
              }
            ],
            "title": "Imagem"
          }
        },
        "type": "object",
        "required": [
          "nome"
        ],
        "title": "PerfilRequest"
      },
...
      "RecuperarSenhaRequest": {
        "properties": {
          "identificador": {
            "type": "string",
            "maxLength": 254,
            "minLength": 2,
            "title": "Identificador"
          }
        },
        "type": "object",
        "required": [
          "identificador"
        ],
        "title": "RecuperarSenhaRequest"
      },
      "RedefinirSenhaRequest": {
        "properties": {
          "identificador": {
            "type": "string",
            "maxLength": 254,
            "minLength": 2,
            "title": "Identificador"
          },
          "codigo": {
            "type": "string",
            "maxLength": 12,
            "minLength": 4,
            "title": "Codigo"
          },
          "nova_senha": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Nova Senha"
          }
        },
        "type": "object",
        "required": [
          "identificador",
          "codigo",
          "nova_senha"
        ],
        "title": "RedefinirSenhaRequest"
      },
...
      "SenhaRequest": {
        "properties": {
          "senha_atual": {
            "type": "string",
            "title": "Senha Atual"
          },
          "nova_senha": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Nova Senha"
          }
        },
        "type": "object",
        "required": [
          "senha_atual",
          "nova_senha"
        ],
        "title": "SenhaRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}